Move code to main module
I prefer to work in files that aren't called `__init__.py`
