+++ pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "listo"
-version = "0.5.0"
+version = "0.6.0"
 description = "An improved version of the List type"
 readme = "README.md"
 authors = [
